CodonMolecularEvolution

Documentation for CodonMolecularEvolution.

CodonMolecularEvolution.benchmark_global_fit

CodonMolecularEvolution.benchmark_grid

CodonMolecularEvolution.difFUBAR_init

CodonMolecularEvolution.getpuresubclades

CodonMolecularEvolution.import_colored_figtree_nexus_as_tagged_tree

CodonMolecularEvolution.optimize_MG94_F3x4

CodonMolecularEvolution.benchmark_global_fit — Method

CodonMolecularEvolution.benchmark_global_fit(benchmark_name; exports=true, data=1:5, optimize_branch_lengths=false)

Benchmarks different implementations of the difFUBARglobalfit algorithm. Results of the benchmark are printed out as a DataFrame and saved to a CSV file. Uses the heuristic top pick to generate con lik matrices. Compares difference in con lik matrices. If exports is true, it also runs MCMCs on the con lik matrices and plots the means and posteriors of the different versions.

benchmark_name is the filepath to where the benchmark will be saved, if exports.

data is the range/vector of datasets to run the benchmark on. By default, this is 1:5. These are the enumerated datasets:

Ace2nobackground

Ace2reallytiny

Ace2tiny

ParvoVP

ParvoVPregrouped

optimize_branch_lengths is an option that can be either true, false or "detect"

CodonMolecularEvolution.benchmark_grid — Method

CodonMolecularEvolution.benchmark_grid(benchmark_name; exports=true, versions_option=1, t::Integer=0, data=1:5)

Benchmarks different implementations of the difFUBAR_grid algorithm. Results of the benchmark are printed out as a DataFrame and saved to a CSV file.

benchmark_name is the filepath to where the benchmark will be saved, if exports

versions_option have 4 different options:

default option, only run heuristic top pick

only run baseline version

run heuristic top pick and baseline version

run all versions

t is the number of threads you want to use in the parallel versions. If specified and non-zero, this will override the number of threads chosen by the heuristic.

data is the range/vector of datasets to run the benchmark on. By default, this is 1:5. These are the enumerated datasets:

Ace2nobackground

Ace2reallytiny

Ace2tiny

ParvoVP

ParvoVPregrouped

CodonMolecularEvolution.difFUBAR_init — Method

CodonMolecularEvolution.getpuresubclades — Method

getpuresubclades(tree::FelNode, tags::Vector{String})

Should usually be called on the root of the tree. Traverses the tree iteratively with a depth-first search to find roots of pure subclades, presuming that nodenames have been trailed with tags. Returns a Vector{FelNode} with root-nodes of the pure subclades.

CodonMolecularEvolution.import_colored_figtree_nexus_as_tagged_tree — Method

scrape_figtree_colored_nexus(fname; custom_labels = String[])

Takes a nexus file from FigTree, where branches have been colored. Replaces all color tags with group tags that can be used in the models. Can add custom labels too. Should consider an entire custom dictionary as well in future.

CodonMolecularEvolution.optimize_MG94_F3x4 — Method

optimize_MG94_F3x4(seqnames, seqs, treestring; leaf_name_transform = x -> x)

Optimizes the MG94+F3x4 model on a tree, given a set of sequences and a tree. Returns the optimized tree, alpha, beta, nucmatrix, F3x4, and eqfreqs. The leafnametransform kwarg can be used to transform the leaf names in the tree to match the seqnames.