CodonMolecularEvolution

Documentation for CodonMolecularEvolution.

Descendant of MolecularEvolution.jl, specializing in codon models.

A Julia package that contains a collection of popular/(state of the art) codon model methods.

CodonMolecularEvolution.jl leverages the MolecularEvolution.jl framework to provide a performant interface to some codon model methods that are used for phylogenetic analysis.

Collection of codon model methods

difFUBAR: Scalable Bayesian comparison of selection pressure

Perform a site-wise comparison of evolutionary pressure between two selected sets of branches.

Design principles

Specificity

We keep our implementations on a high level of abstraction and let MolecularEvolution.jl take care of the low-level behaviour.

Scalability

Our exported codon model methods should scale to large, real-world datasets. To keep the memory footprint down, we use MolecularEvolution.jls LazyPartition when possible.

Performance

While scalability takes precedence over speed?, we try to maintain competitive runtimes by using e.g. computational shortcuts and parallelization whenever we can.

Authors

Ben decides what to put here.

CodonMolecularEvolution.difFUBARBaseline

CodonMolecularEvolution.difFUBARParallel

CodonMolecularEvolution.difFUBARTreesurgery

CodonMolecularEvolution.difFUBARTreesurgeryAndParallel

CodonMolecularEvolution.benchmark_global_fit

CodonMolecularEvolution.benchmark_grid

CodonMolecularEvolution.difFUBAR

CodonMolecularEvolution.difFUBAR

CodonMolecularEvolution.difFUBAR_init

CodonMolecularEvolution.getpuresubclades

CodonMolecularEvolution.import_colored_figtree_nexus_as_tagged_tree

CodonMolecularEvolution.import_colored_figtree_nexus_as_tagged_tree

CodonMolecularEvolution.import_grouped_label_tree

CodonMolecularEvolution.import_grouped_label_tree

CodonMolecularEvolution.optimize_MG94_F3x4

CodonMolecularEvolution.optimize_MG94_F3x4

CodonMolecularEvolution.benchmark_global_fit — Method

CodonMolecularEvolution.benchmark_global_fit(benchmark_name; exports=true, data=1:5, optimize_branch_lengths=false)

Benchmarks different implementations of the difFUBARglobalfit algorithm. Results of the benchmark are printed out as a DataFrame and saved to a CSV file. Uses the heuristic top pick to generate con lik matrices. Compares difference in con lik matrices. If exports is true, it also runs MCMCs on the con lik matrices and plots the means and posteriors of the different versions.

benchmark_name is the filepath to where the benchmark will be saved, if exports.

data is the range/vector of datasets to run the benchmark on. By default, this is 1:5. These are the enumerated datasets:

Ace2nobackground

Ace2reallytiny

Ace2tiny

ParvoVP

ParvoVPregrouped

optimize_branch_lengths is an option that can be either true, false or "detect"

CodonMolecularEvolution.benchmark_grid — Method

CodonMolecularEvolution.benchmark_grid(benchmark_name; exports=true, versions_option=1, t::Integer=0, data=1:5)

Benchmarks different implementations of the difFUBAR_grid algorithm. Results of the benchmark are printed out as a DataFrame and saved to a CSV file.

benchmark_name is the filepath to where the benchmark will be saved, if exports

versions_option have 4 different options:

default option, only run heuristic top pick

only run baseline version

run heuristic top pick and baseline version

run all versions

t is the number of threads you want to use in the parallel versions. If specified and non-zero, this will override the number of threads chosen by the heuristic.

data is the range/vector of datasets to run the benchmark on. By default, this is 1:5. These are the enumerated datasets:

Ace2nobackground

Ace2reallytiny

Ace2tiny

ParvoVP

ParvoVPregrouped

CodonMolecularEvolution.difFUBAR — Method

difFUBAR(seqnames, seqs, treestring, tags, tag_colors, outpath; <keyword arguments>)

Takes a tagged phylogeny and an alignment as input and performs difFUBAR analysis.

Arguments

seqnames: vector of untagged sequence names.

seqs: vector of aligned sequences, corresponding to seqnames.

treestring: a tagged newick tree string.

tags: vector of tag signatures.

tag_colors: vector of tag colors (hex format).

outpath: export directory.

pos_thresh=0.95: threshold of significance for the posteriors.

iters=2500: iterations used in the Gibbs sampler.

verbosity=1: as verbosity increases, prints are added accumulatively.

0 - no prints

1 - show current step and where output files are exported

2 - show the chosen difFUBAR_grid version and amount of parallel threads.

exports=true: if true, output files are exported.

code=MolecularEvolution.universal_code: genetic code used for the analysis.

optimize_branch_lengths=false: if true, the branch lengths of the phylogenetic tree are optimized.

version::Union{difFUBARGrid, Nothing}=nothing: explicitly choose the version of difFUBAR_grid to use. If nothing, the version is heuristically chosen based on the available RAM and Julia threads.

t=0: explicitly choose the amount of Julia threads to use. If 0, the degree of parallelization is heuristically chosen based on the available RAM and Julia threads.

Note

Julia starts up with a single thread of execution, by default. See Starting Julia with multiple threads.

CodonMolecularEvolution.difFUBAR_init — Method

CodonMolecularEvolution.getpuresubclades — Method

getpuresubclades(tree::FelNode, tags::Vector{String})

Should usually be called on the root of the tree. Traverses the tree iteratively with a depth-first search to find roots of pure subclades, presuming that nodenames have been trailed with tags. Returns a Vector{FelNode} with root-nodes of the pure subclades.

CodonMolecularEvolution.import_colored_figtree_nexus_as_tagged_tree — Method

import_colored_figtree_nexus_as_tagged_tree(fname; custom_labels=String[])

Takes a nexus file from FigTree, where branches have been colored. Replaces all color tags with group tags that can be used in the models. Can add custom labels too. Should consider an entire custom dictionary as well in future.

Examples

julia> treestring, tags, tag_colors = import_colored_figtree_nexus_as_tagged_tree("data/Ace2_no_background.nex")
("(((XM_027533928_Bos_indicus_x_Bos_taurus{G1}:0.097072,(XM_042974087_Panthera_tigris{G1}:0.038016,... more ...;", ["{G2}", "{G1}"], ["#ff0015", "#0011ff"])

Note

treestring is truncated. NEXUS tree file

CodonMolecularEvolution.import_grouped_label_tree — Method

import_grouped_label_tree(tree_file)

Takes a Newick tree file and return Newick tree, Newick tree with replaced tags, group tags, original tags, and randomly generated colours for each tag

CodonMolecularEvolution.optimize_MG94_F3x4 — Method

optimize_MG94_F3x4(seqnames, seqs, tree; leaf_name_transform=x -> x, genetic_code=MolecularEvolution.universal_code)

Optimizes the MG94+F3x4 model on a tree, given a set of sequences and a tree. Returns the optimized tree, alpha, beta, nucmatrix, F3x4, and eqfreqs. The leafnametransform kwarg can be used to transform the leaf names in the tree to match the seqnames.